+++ NEWS.md
@@ -1,6 +1,8 @@
 # crew 0.10.2
 
-* Eliminate spurious `launch_max` error from underutilized workers and deprecate `launch_max` in favor of `crashes_error` (#189).
+* Eliminate spurious `launch_max` error from underutilized workers (#189).
+* Deprecate `launch_max` in favor of `crashes_error` (#189).
+* Look for crashes of all workers in `rotate()` instead of looking for crashes of a specific worker in `launch()` (#189).
 
 # crew 0.10.1